+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v1.5.2
+- Reimplemented argv management with minimist
+- Now readJson makes use of require instead of readFileSync
+- Written help function
+- Changed license, moved form GPLv2 to MIT
+
 ## v1.5.0
 - Added automatic Git tag
 - 'apv init' uses the version of package.json when creates the appversion.json file.